requirements.txt
jinja2>=3.1.4
pyyaml>=6.0.1
google-cloud-aiplatform
ipykernel